@@ -11,14 +11,10 @@
     "test": "vitest"
   },
   "dependencies": {
-    "@event-calendar/day-grid": "^2.5.1",
-    "@event-calendar/interaction": "^2.5.1",
-    "@event-calendar/list": "^2.5.1",
-    "@event-calendar/resource-time-grid": "^2.5.1",
-    "@event-calendar/time-grid": "^2.5.1",
     "@floating-ui/dom": "^1.5.3",
     "@supabase/supabase-js": "^2.38.1",
     "@tauri-apps/api": "^2.0.0-rc.0",
+    "@tauri-apps/plugin-clipboard-manager": "^2.0.0-rc.1",
     "@tauri-apps/plugin-dialog": "^2.0.0-rc.0",
     "@tauri-apps/plugin-fs": "^2.0.0-rc.0",
     "@tauri-apps/plugin-http": "^2.0.0-rc.0",
@@ -26,6 +22,7 @@
     "@tauri-apps/plugin-process": "^2.0.0-rc.0",
     "@tauri-apps/plugin-shell": "^2.0.0-rc.0",
     "@tauri-apps/plugin-sql": "^2.0.0-rc.0",
+    "@tauri-apps/plugin-stronghold": "^2.0.0-rc.1",
     "@tauri-apps/plugin-updater": "^2.0.0-rc.0",
     "dayjs": "^1.11.10",
     "jspdf": "^2.5.1",
@@ -33,24 +30,33 @@
     "svelte-persisted-store": "^0.7.0"
   },
   "devDependencies": {
-    "@skeletonlabs/skeleton": "^2.6.1",
-    "@skeletonlabs/tw-plugin": "^0.3.0",
-    "@sveltejs/adapter-static": "^2.0.3",
-    "@sveltejs/kit": "^1.25.2",
-    "@sveltejs/vite-plugin-svelte": "^2.4.2",
-    "@tailwindcss/forms": "^0.5.6",
+    "@event-calendar/core": "^3.7.1",
+    "@event-calendar/day-grid": "^3.7.1",
+    "@event-calendar/interaction": "^3.7.1",
+    "@event-calendar/list": "^3.7.1",
+    "@event-calendar/resource-time-grid": "^3.7.1",
+    "@event-calendar/time-grid": "^3.7.1",
+    "@skeletonlabs/skeleton": "^2.10.3",
+    "@skeletonlabs/tw-plugin": "^0.4.0",
+    "@sveltejs/adapter-static": "^3.0.6",
+    "@sveltejs/kit": "^2.7.3",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
+    "@tailwindcss/forms": "^0.5.9",
+    "@tailwindcss/typography": "^0.5.15",
     "@tauri-apps/cli": "^2.0.0-rc.3",
     "autoprefixer": "^10.4.16",
-    "eslint": "^8.51.0",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-svelte": "^2.36.0-next.13",
     "internal-ip": "^7.0.0",
-    "postcss": "^8.4.31",
-    "prettier": "^3.0.3",
-    "prettier-plugin-svelte": "^3.0.3",
-    "prettier-plugin-tailwindcss": "^0.5.6",
-    "svelte": "^4.2.8",
-    "tailwindcss": "^3.3.3",
-    "vite": "^4.4.4",
-    "vite-plugin-tailwind-purgecss": "^0.1.3",
+    "postcss": "^8.4.38",
+    "prettier": "^3.2.5",
+    "prettier-plugin-svelte": "^3.1.2",
+    "prettier-plugin-tailwindcss": "^0.5.14",
+    "svelte": "^5.1.3",
+    "tailwindcss": "^3.4.3",
+    "vite": "^5.0.3",
+    "vite-plugin-tailwind-purgecss": "^0.3.3",
     "vitest": "^1.0.4"
   }
 }
