@@ -1,105 +1,81 @@
 <script>
 	import PasswordField from './PasswordField.svelte';
 	import { supabase } from '../../stores/supabaseClient';
+	import { initializeConventions } from '../../stores/conventionStore';
 	import { goto } from '$app/navigation';
 	import { user } from '$lib/stores/UserStore';
 	import { patients } from '../../stores/PatientStore';
 	import EmailField from './EmailField.svelte';
 	import { SubmitButton, FormWrapper } from '../index';
-	import { DBInitializer } from '../../stores/databaseInitializer';
-	import DBAdapter from '../actions/dbAdapter';
-	import dayjs from 'dayjs';
+	import { LocalDatabase } from '../../stores/databaseInitializer';
+	import DBAdapter from '$lib/user-ops-handlers/dbAdapter';
 	import { get } from 'svelte/store';
 	import { t } from '../../i18n';
+	import { file_exists } from '../../utils/fsAccessor';
+	import { tick } from 'svelte';
+	import { retrieveSettings } from '../../user-ops-handlers/settings';
+	import { listPatients } from '../../user-ops-handlers/patients';
+	import { signUserIn } from '../../user-ops-handlers/users';
 
 	export let message = '';
 
 	// <!--* Login validation -->
 	async function signIn({ formData, submitter }) {
 		// <!--* STEP 1 : Connecter l'utilisateur -->
-		const { data, error } = await supabase.auth.signInWithPassword({
-			email: formData.email.toLowerCase(),
-			password: formData.password
-		});
-		let db = new DBAdapter();
-		// <!--? STEP 1a : Error handling -->
-		if (error) {
-			message = error.message;
-			submitter.disabled = false;
-			spinner.classList.remove('!block');
-			return;
-		}
-		user.set({
-			user: data.user,
-			session: data.session
-		});
-		message = `${get(t)('shared', 'welcome')} ${data.user.email}!`;
+		await signUserIn(formData);
+		message = `${get(t)('shared', 'welcome')} ${get(user).email}!`;
 		console.log('user = ', $user);
 		submitter.innerHTML = get(t)('login', 'submission.profil');
 		// <!--* STEP 2 : Initialiser la DB -->
 		// <!--? Télécharger et mettre à jour les codes de nomenclature -->
 		submitter.innerHTML = get(t)('login', 'submission.convention');
-		let dbInit = new DBInitializer();
-		await dbInit.initialization(submitter);
-		// <!--* STEP 3 : Récupérer les données de supabase -->
-		// <!--? pour l'instant c'est inutile car on ne garde que le nom et prénom sur le serveur -->
-		let kineData = await supabase.from('kinesitherapeutes').select().eq('id', data.user.id);
-		// <!--? Si les données n'existe pas on va pas plus loin, on envoie sur post-signup-form -->
-		if (kineData.data.length === 0) {
-			submitter.disabled = false;
-			goto('/post-signup-forms/kine-profile');
-			return;
-		}
+		await initializeConventions(submitter);
 		// <!--* STEP 4 : Récupérer les données de la DB -->
-		// <!--? Normalement, si il y a un profil sur le serveur, il devrait y en avoir un sur la db -->
-		let kineData2 = await db.retrieve('kines', '*', ['user_id', data.user.id]);
-		console.log('kineData2', kineData2);
-		// <!--? Si les données n'existe pas on va pas plus loin, on envoie sur post-signup-form -->
-		if (kineData2.data.length === 0) {
+		// <!--* STEP 8 : Setup la sécurité -->
+		// le cas de figure le plus fort c'est qu'il n'y a pas de stronghold key
+		// cela signifie que rien n'est mis en place en ce qui concerne la sécurité
+		// On redirige donc vers la page de mise en place de la sécurité
+		submitter.innerHTML = get(t)('login', 'submission.security', null, 'Security check');
+		await retrieveSettings({ user_id: $user.user.id });
+		if (!$user.profil.has_stronghold_key) {
+			goto('/post-signup-forms/encryption-key-setup');
+			return;
+		}
+		// si il y a une stronghold key il faut vérifier qu'il y a un stronghold
+		// avec une clé d'encryption enregestrée dessus
+		// par chance, la fonction getMainKey() contient la fonction initStronghold()
+		// qui elle s'assure d'elle-même de la présence d'un fichier salt.txt sur applocaldatadir(),
+		// ouvre ou crée un stronghold et
+		// tente d'aller y puiser la clé de cryptage.
+		// Si getMainKey retourne null, alors on envoie sur la page de setup avec un query parameter pour proposer en premier le formulaire plutôt que générer une clé d'encryption
+		/**
+		 * TODO Wrap this in userOpsHandler
+		 * */
+		let hold_exists = await file_exists(`${$user.user.id}.hold`);
+		console.log('file exists', hold_exists);
+
+		if (!hold_exists) {
+			goto('/post-signup-forms/encryption-key-setup?cloud=true');
+			return;
+		}
+
+		// <!--* STEP 5 : Fusionner les données de supabase et de la db -->
+		if (!user.has_complete_profile()) {
 			submitter.disabled = false;
 			goto('/post-signup-forms/kine-profile');
 			return;
 		}
-		// <!--* STEP 5 : Fusionner les données de supabase et de la db -->
-		kineData = {
-			user_id: kineData.data[0].id,
-			nom: kineData.data[0].nom,
-			prenom: kineData.data[0].prenom,
-			inami: kineData2.data[0].inami,
-			bce: kineData2.data[0].bce,
-			iban: kineData2.data[0].iban,
-			adresse: kineData2.data[0].adresse,
-			cp: kineData2.data[0].cp,
-			localite: kineData2.data[0].localite,
-			conventionne: kineData2.data[0].conventionne
-		};
-		user.update((u) => {
-			u.profil = kineData;
-			return u;
-		});
 		submitter.innerHTML = get(t)('login', 'submission.patient');
 		// <!--* STEP 6 : Récupérer les patients -->
-		await patients.fetchPatient(data.user);
-		submitter.innerHTML = get(t)('login', 'submission.settings');
+		console.log('Now preparing to fetch patients with user = ', $user);
+
+		await listPatients(get(user).user);
+		submitter.innerHTML = get(t)('login', 'submission.settings', null, 'Gathering settings');
 		// <!--* STEP 7 : Récupérer les settings -->
-		let userSettings = await db.retrieve('settings', '*', ['user_id', $user.user.id]);
-		console.log('userSettings', userSettings);
-		if (userSettings.data.length === 0) {
-			await db.save('settings', {
-				user_id: $user.user.id,
-				created_at: dayjs().format('YYYY-MM-DD HH:mm:ss')
-			});
-			userSettings.data[0] = { raw_printer: null };
-		}
-		userSettings = userSettings.data[0];
-		userSettings.is_nine_pin = userSettings.is_nine_pin === null ? false : JSON.parse(userSettings.is_nine_pin);
-		// <!--* STEP 8 : Sauvegarder les données dans le store -->
-		user.update((u) => {
-			u.settings = userSettings;
-			return u;
-		});
+		await retrieveSettings({ user_id: $user.user.id });
+
 		// <!--* STEP 9 : Si il n'y a pas d'imprimante matricielle on redirige vers la page setup correspondante -->
-		if (!userSettings.raw_printer) {
+		if (!$user.settings.raw_printer) {
 			goto('/post-signup-forms/printer-setup');
 			return;
 		}
