@@ -1,15 +1,16 @@
-import adapter from '@sveltejs/adapter-static' // This was changed from adapter-auto
-import { vitePreprocess } from '@sveltejs/kit/vite'
+import adapter from '@sveltejs/adapter-static'; // This was changed from adapter-auto
+import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
+
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
-  // Consult https://kit.svelte.dev/docs/integrations#preprocessors
-  // for more information about preprocessors
-  preprocess: vitePreprocess(),
+	// Consult https://kit.svelte.dev/docs/integrations#preprocessors
+	// for more information about preprocessors
+	preprocess: vitePreprocess(),
 
-  kit: {
-    adapter: adapter(),
-  },
-}
+	kit: {
+		adapter: adapter()
+	}
+};
 
-export default config
+export default config;
